+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Return plain text errors for routes other than Mastodon API.
 - Changed MSRV to 1.72.1.
 - Don't reject actor if associated public key has different origin.
+- Set `Accept` header value to `application/jrd+json` when performing WebFinger query.
 
 ## [4.3.1] - 2025-05-18